refactor : style 위치 변경
