+++ src/components/chat/ChatHeader.tsx
@@ -1,9 +1,9 @@
 import {IoChevronBackOutline, IoCloseOutline} from 'react-icons/io5';
-import {Header, BackButton, UserName, CloseButton} from '../../styles/style.chatheader';
 import {showDivState, hideButtonState, userInfo, partnerInfo} from '../../store/atom';
 import {useRecoilState ,useSetRecoilState} from 'recoil';
 import {IUser} from '../../store/interface';
 import {useNavigate} from "react-router-dom";
+import styled from 'styled-components';
 
 function ChatHeader(): JSX.Element {
   const setShowDiv = useSetRecoilState<boolean>(showDivState);
@@ -35,3 +35,32 @@ function ChatHeader(): JSX.Element {
 }
     
 export default ChatHeader;
+
+const Header = styled.div`
+    display : flex;
+    align-items : center;
+    padding : 10px;
+    background-color: rgba(239, 239, 240, 0.85);
+    margin-bottom: 5px;
+`
+
+const BackButton = styled.div`
+    display: flex;
+    align-items: center;
+    min-width: 25px;
+    padding: 10px 0px;
+    margin-right: 15px;
+    cursor: pointer;
+`
+
+const UserName = styled.span`
+    width : 100%;
+    font-size : 1.3rem;
+    font-style : normal;
+    font-weight : bold;
+    user-select : none;
+    cursor : pointer;
+`
+const CloseButton = styled.div`
+    cursor : pointer;
+`
