+++ src/components/chat/ChatListComponent.tsx
@@ -56,9 +56,9 @@ export default ChatList;
 const ChatListWrapper = styled.div.attrs({
   tabIndex: 0,
 })`
-	height : 100%;
+  height : 100%;
   padding : 1rem;
-  overflow : scroll;
+  overflow : auto;
   &.show{
     margin-top : 50px;
   }
