+++ src/components/friends/FriendList.tsx
@@ -46,9 +46,9 @@ export default FriendList;
 const FriendsWrapper = styled.div.attrs({
   tabIndex: 0,
 })`
-	height : 100%;
+  height : 100%;
   padding : 1rem;
-  overflow : scroll;
+  overflow : auto;
   &.show{
     margin-top : 50px;
   }
