--- src/styles/style.chatinput.ts
@@ -1,25 +1,0 @@
-import styled from 'styled-components';
-
-export const ChatInputWrapper = styled.form`
-    margin : 0px 10px 10px;
-    display : flex;
-`
-
-export const Input = styled.textarea`
-    padding: 10px;
-    width:80%;
-    border: none;
-    font-size: 1.2rem;
-    line-height: 1.5rem;
-    outline: none;
-    resize: none;
-    background-color:transparent;
-`
-
-export const SubmitButton = styled.button`
-    padding: 10px;
-    font-size: 1.2rem;
-    cursor: pointer;
-    border : none;
-    background-color:transparent;
-`
